fix(core): restore ares-v3 build and clean bench corpus fixture
Pin printpdf to 0.7.0 — the 0.12 bump broke PdfDocument::new/use_text APIs.
Adjust font sizes to f32 and map ratatui draw errors for the 0.30 backend.
Remove duplicate Google-key line in mobile-static corpus.json (invalid JSON
left by a partial secret-scanning merge).

+++ core/crates/ares-cli/src/commands/pdf.rs
@@ -56,7 +56,7 @@ fn build_pdf(json: &Value, output: &Path) -> AresResult<()> {
         x: Mm,
         y: &mut Mm,
         font: &IndirectFontRef,
-        size: f64,
+        size: f32,
     ) {
         layer.use_text(text, size, x, *y, font);
         *y = Mm(y.0 - size * 0.4);
@@ -382,7 +382,7 @@ pub fn generate_scan_pdf(report: &ares_core::AuditReport, output: &Path) -> Ares
         x: Mm,
         y: &mut Mm,
         font: &IndirectFontRef,
-        size: f64,
+        size: f32,
     ) {
         layer.use_text(text, size, x, *y, font);
         *y = Mm(y.0 - size * 0.4);
